updated int

+++ cymetric/pointgen/pointgen.py
@@ -86,10 +86,10 @@ class PointGenerator:
         else:
             level = logging.WARNING
         logger.setLevel(level=level)
-        self.monomials = monomials.astype(int64)
+        self.monomials = monomials.astype(int)
         self.coefficients = coefficients
         self.kmoduli = kmoduli
-        self.ambient = ambient.astype(int64)
+        self.ambient = ambient.astype(int)
         self.degrees = ambient + 1
         self.nhyper = 1
         self.nmonomials, self.ncoords = monomials.shape
@@ -128,7 +128,7 @@ class PointGenerator:
         needed for point generation, residue theorem
         and the pullback tensor.
         """
-        self.all_ts = np.eye(len(self.ambient), dtype=int64)
+        self.all_ts = np.eye(len(self.ambient), dtype=int)
         self.selected_t = self.all_ts[np.argmax(self.ambient)]
         self._generate_root_basis()
         self._generate_dQdz_basis()
@@ -812,8 +812,8 @@ class PointGenerator:
             shapes = np.array([[[np.shape(t[0]), np.shape(t[1])]
                                 if i != j else [[-1, -1], [-1, -1]] for i, t in enumerate(zi)]
                                for j, zi in enumerate(self.dzdz_basis)])
-            DZDZB_d = np.zeros((len(shapes), len(shapes), np.max(shapes[:, :, 0, 0]), len(shapes)), dtype=int64)
-            DZDZB_n = np.zeros((len(shapes), len(shapes), np.max(shapes[:, :, 1, 0]), len(shapes)), dtype=int64)
+            DZDZB_d = np.zeros((len(shapes), len(shapes), np.max(shapes[:, :, 0, 0]), len(shapes)), dtype=int)
+            DZDZB_n = np.zeros((len(shapes), len(shapes), np.max(shapes[:, :, 1, 0]), len(shapes)), dtype=int)
             DZDZF_d = np.zeros((len(shapes), len(shapes), np.max(shapes[:, :, 0, 0])), dtype=np.complex128)
             DZDZF_n = np.zeros((len(shapes), len(shapes), np.max(shapes[:, :, 1, 0])), dtype=np.complex128)
             for i in range(len(shapes)):
@@ -1040,7 +1040,7 @@ class PointGenerator:
 
         Args:
             points (ndarray[(n_p, ncoords), np.complex128]): Points.
-            j_elim (ndarray([n_p], int64)): index to be eliminated. 
+            j_elim (ndarray([n_p], int)): index to be eliminated. 
                 Defaults not None. If None eliminates max(dQdz).
 
         Returns:
@@ -1061,7 +1061,7 @@ class PointGenerator:
             points (ndarray[(n_p, ncoords), np.complex128]): Points.
 
         Returns:
-            ndarray[(n_p), int64]: max(dQdz) indices
+            ndarray[(n_p), int]: max(dQdz) indices
         """
         dQdz = np.abs(self._compute_dQdz(points))
         dQdz = dQdz * (~np.isclose(points, complex(1, 0)))
@@ -1123,7 +1123,7 @@ class PointGenerator:
                                 &= d^{ijk} t_i t_j t_k.
 
                 Defaults to False.
-            j_elim (ndarray([n_p, nhyper], int64)): Index to be eliminated. 
+            j_elim (ndarray([n_p, nhyper], int)): Index to be eliminated. 
                 Defaults to None. If None eliminates max(dQdz).
 
         Returns:
@@ -1189,7 +1189,7 @@ class PointGenerator:
 
         Args:
             points (ndarray([n_p, ncoords], np.complex128)): Points.
-            j_elim (ndarray([n_p, nhyper], int64)): Index to be eliminated. 
+            j_elim (ndarray([n_p, nhyper], int)): Index to be eliminated. 
                 Defaults to None. If None eliminates max(dQdz).
 
         Returns:
